separated login with home screen
