+++ miniproj/src/firebase.js
@@ -1,6 +1,7 @@
-import {initializeApp} from 'firebase/app';
+//import {initializeApp} from 'firebase/app';
 import {getAuth, GoogleAuthProvider} from 'firebase/auth';
 import firebase from 'firebase/compat/app';
+import { getFirestore } from 'firebase/firestore';
 import 'firebase/compat/firestore';
 
 const firebaseConfig = {
